missing change for last commit
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@6585 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/net/yacy/kelondro/table/Table.java
@@ -284,8 +284,8 @@ public class Table implements ObjectIndex, Iterable<Row.Entry> {
         return this.table != null;
     }
     
-    public static int staticRAMIndexNeed(final File f, final Row rowdef) throws IOException {
-        return (int) (((long)(rowdef.primaryKeyLength + 4)) * tableSize(f, rowdef.objectsize) * RowCollection.growfactorLarge100 / 100L);
+    public static long staticRAMIndexNeed(final File f, final Row rowdef) throws IOException {
+        return (((long)(rowdef.primaryKeyLength + 4)) * tableSize(f, rowdef.objectsize) * RowCollection.growfactorLarge100 / 100L);
     }
     
     public synchronized void addUnique(final Entry row) throws IOException, RowSpaceExceededException {
